Reorganize pipeline configuration files
